Core: Moved and exported cell system.
