+++ plugins/org.eclipse.elk.core/META-INF/MANIFEST.MF
@@ -19,6 +19,7 @@ Export-Package: org.eclipse.elk.core,
  org.eclipse.elk.core.util.adapters,
  org.eclipse.elk.core.util.internal;x-friends:="org.eclipse.elk.graph.text",
  org.eclipse.elk.core.util.nodespacing,
+ org.eclipse.elk.core.util.nodespacing.cellsystem,
  org.eclipse.elk.core.util.persistence,
  org.eclipse.elk.core.util.selection
 Bundle-Vendor: Eclipse Modeling Project
