+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/nodespacing/cellsystem/AtomicCell.java
@@ -8,7 +8,7 @@
  * Contributors:
  *    Kiel University - initial API and implementation
  *******************************************************************************/
-package org.eclipse.elk.core.util.nodespacing.internal.cellsystem;
+package org.eclipse.elk.core.util.nodespacing.cellsystem;
 
 import org.eclipse.elk.core.math.ElkPadding;
 import org.eclipse.elk.core.math.KVector;
