+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/nodespacing/cellsystem/HorizontalLabelAlignment.java
@@ -8,18 +8,18 @@
  * Contributors:
  *     Kiel University - initial API and implementation
  *******************************************************************************/
-package org.eclipse.elk.core.util.nodespacing.internal;
+package org.eclipse.elk.core.util.nodespacing.cellsystem;
 
 /**
- * Vertical alignment of labels.
+ * Horizontal alignment of labels.
  */
-public enum VerticalLabelAlignment {
+public enum HorizontalLabelAlignment {
     
-    /** Labels are top-aligned. */
-    TOP,
+    /** Labels are left-aligned. */
+    LEFT,
     /** Labels are centered. */
     CENTER,
-    /** Labels are bottom-aligned. */
-    BOTTOM;
+    /** Labels are right-aligned. */
+    RIGHT;
     
 }
