+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/nodespacing/cellsystem/StripContainerCell.java
@@ -8,7 +8,7 @@
  * Contributors:
  *    Kiel University - initial API and implementation
  *******************************************************************************/
-package org.eclipse.elk.core.util.nodespacing.internal.cellsystem;
+package org.eclipse.elk.core.util.nodespacing.cellsystem;
 
 import java.util.Arrays;
