+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/nodespacing/internal/PortContext.java
@@ -12,7 +12,7 @@ package org.eclipse.elk.core.util.nodespacing.internal;
 
 import org.eclipse.elk.core.math.ElkMargin;
 import org.eclipse.elk.core.util.adapters.GraphAdapters.PortAdapter;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.LabelCell;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.LabelCell;
 
 /**
  * Data holder class to be passed around to avoid having too much state in the size calculation classes. Port contexts
