+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/nodespacing/internal/algorithm/InsidePortLabelCellCreator.java
@@ -14,11 +14,11 @@ import org.eclipse.elk.core.math.ElkPadding;
 import org.eclipse.elk.core.math.KVector;
 import org.eclipse.elk.core.options.PortLabelPlacement;
 import org.eclipse.elk.core.options.PortSide;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.AtomicCell;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.ContainerArea;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.StripContainerCell;
 import org.eclipse.elk.core.util.nodespacing.internal.NodeContext;
 import org.eclipse.elk.core.util.nodespacing.internal.PortContext;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.AtomicCell;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.ContainerArea;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.StripContainerCell;
 
 /**
  * Sets up the inside port label cells. These are set up even when there are no inside port labels since they also
