+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/nodespacing/internal/algorithm/NodeLabelCellCreator.java
@@ -17,13 +17,13 @@ import org.eclipse.elk.core.options.NodeLabelPlacement;
 import org.eclipse.elk.core.options.PortSide;
 import org.eclipse.elk.core.options.SizeOptions;
 import org.eclipse.elk.core.util.adapters.GraphAdapters.LabelAdapter;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.ContainerArea;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.GridContainerCell;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.LabelCell;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.StripContainerCell;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.StripContainerCell.Strip;
 import org.eclipse.elk.core.util.nodespacing.internal.NodeContext;
 import org.eclipse.elk.core.util.nodespacing.internal.NodeLabelLocation;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.ContainerArea;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.GridContainerCell;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.LabelCell;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.StripContainerCell;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.StripContainerCell.Strip;
 
 /**
  * Knows how to take all of a node's labels and create the appropriate grid cells.
