+++ plugins/org.eclipse.elk.core/src/org/eclipse/elk/core/util/nodespacing/internal/algorithm/VerticalPortPlacementSizeCalculator.java
@@ -17,9 +17,9 @@ import org.eclipse.elk.core.options.PortLabelPlacement;
 import org.eclipse.elk.core.options.PortSide;
 import org.eclipse.elk.core.options.SizeConstraint;
 import org.eclipse.elk.core.options.SizeOptions;
+import org.eclipse.elk.core.util.nodespacing.cellsystem.AtomicCell;
 import org.eclipse.elk.core.util.nodespacing.internal.NodeContext;
 import org.eclipse.elk.core.util.nodespacing.internal.PortContext;
-import org.eclipse.elk.core.util.nodespacing.internal.cellsystem.AtomicCell;
 
 /**
  * Calculates the space required to setup port labels.
